refactor : Card --> shared Layer
