+++ src/components/index.tsx
@@ -4,74 +4,6 @@ import * as SelectPrimitive from '@radix-ui/react-select'
 import * as DialogPrimitive from '@radix-ui/react-dialog'
 import { Check, ChevronDown, X } from 'lucide-react'
 
-// 입력 컴포넌트
-interface InputProps extends React.InputHTMLAttributes<HTMLInputElement> {
-  className?: string
-}
-
-export const Input = forwardRef<HTMLInputElement, InputProps>(
-  ({ className, type, ...props }, ref) => {
-    return (
-      <input
-        type={type}
-        className={`flex h-10 w-full rounded-md border border-input bg-white px-3 py-2 text-sm ring-offset-background file:border-0 file:bg-transparent file:text-sm file:font-medium placeholder:text-muted-foreground focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 disabled:cursor-not-allowed disabled:opacity-50 ${className}`}
-        ref={ref}
-        {...props}
-      />
-    )
-  }
-)
-Input.displayName = 'Input'
-
-// 카드 컴포넌트
-interface CardProps extends React.HTMLAttributes<HTMLDivElement> {
-  className?: string
-}
-
-export const Card = forwardRef<HTMLDivElement, CardProps>(
-  ({ className, ...props }, ref) => (
-    <div
-      ref={ref}
-      className={`rounded-lg border bg-card text-card-foreground shadow-sm ${className}`}
-      {...props}
-    />
-  )
-)
-Card.displayName = 'Card'
-
-export const CardHeader = forwardRef<HTMLDivElement, CardProps>(
-  ({ className, ...props }, ref) => (
-    <div
-      ref={ref}
-      className={`flex flex-col space-y-1.5 p-6 ${className}`}
-      {...props}
-    />
-  )
-)
-CardHeader.displayName = 'CardHeader'
-
-interface CardTitleProps extends React.HTMLAttributes<HTMLHeadingElement> {
-  className?: string
-}
-
-export const CardTitle = forwardRef<HTMLHeadingElement, CardTitleProps>(
-  ({ className, ...props }, ref) => (
-    <h3
-      ref={ref}
-      className={`text-2xl font-semibold leading-none tracking-tight ${className}`}
-      {...props}
-    />
-  )
-)
-CardTitle.displayName = 'CardTitle'
-
-export const CardContent = forwardRef<HTMLDivElement, CardProps>(
-  ({ className, ...props }, ref) => (
-    <div ref={ref} className={`p-6 pt-0 ${className}`} {...props} />
-  )
-)
-CardContent.displayName = 'CardContent'
-
 // 텍스트 영역 컴포넌트
 interface TextareaProps
   extends React.TextareaHTMLAttributes<HTMLTextAreaElement> {
