+++ src/shared/index.ts
@@ -1,2 +1,4 @@
 export { Pagination } from './ui/Pagination'
 export { Button } from './ui/Button'
+export { Input } from './ui/Input'
+export { Card, CardHeader, CardTitle, CardContent } from './ui/Card'
